Introducing "graph" and "classic" modes.
A server can support "graph" mode, "classic" mode, or both.

If it supports "classic" mode, it will make sure to fill in all of
the fields that non-graph clients require, and not use incompatible
graph extensions like `GraphAlignment`.

If it supports "graph" mode, it will make sure to fill in all of the new
graph fields, like `Reference.segment` and `Variant.alleles`. Additionally,
if it also does not support "classic" mode, it will not use older, less
general things like `LinearAlignment` and single string phase sets.